Add enhanced article types and fetching mechanism
Introduced new types for articles and authors in the system.
Added utility function to convert content to HTML.
Implemented fetching of versions with no content and prepared authors and
subjects for the system. Also, added a converter from EnhancedArticleNoContent
to ReviewedPreprintSnippet. These changes support the enhancement of
articles in the system.